requirements.txt
aiohttp
aiohttp-sessions
python-socketio
redis